Load modules into parent `namespace` environment
In order to implement `attach`, names must be imported into a separate
namespace environment, and then copy the functions into the exported
module environment. That way, all the names are available but still
inside their own environment, and can themselves access imported names
which are not globally visible.

Furthermore, `name` and `path` should actually be *attributes* of the
namespace, not member objects. This should have been done anyway.