Fix a regression introduced by 5918173
The commit introduced an overload of `module_name` for environments to
make package modules work. However, this breaks the semantic of
`module_name`. A better and simpler way is to use the module name as
it’s already known at this point in the code.

Furthermore, the `import` code still contained some legacy logic
converting the name to a character string, which has been redundant for
a long time.