Failed reloading now restores previous module
If a module fails to load due to an error in the module’s code, `reload`
now restore the previously loaded module and its attached environments
to full function. This fixes #58.

The change also ensures that the *same* module object is overwritten
with the newly loaded module, rather than overwriting a shadowed object
(fixes #51).