Skeleton code for new `mod::use` API
This is still a proof of concept implementation to illustrate the logic
outline. The actual code, especially in the second half of the file, is
subject to change. In particular, this currently fails for circular
module dependencies due to the circular import on the (now) cached first
module, which will attempt to access the not-yet-set module exports.

Furthermore, module loading from inside packages has never been tested.
All other code is still lacking systematic testing.