Correctly implement the new mod environment schema
The schema is explained in `vignettes/mod-env-hierarchy.rmd` on branch
`pkgdown`. This is a first draft. It has not been extensively tested.

Notably, this change removes the automatic loading of “parent” modules.
The rationale for loading these does not form a compelling case,
deviates from other module systems and is thus surprising, and makes the
logic more complicated.