WIP rewrite of the export specification parsing
The rewrite turns out to be quite extensive because it (a) required
changes to lots of parts of the working code, and (b) made lots of other
code simpler as a consequence. Furthermore, testing these changes
required adapting many unit tests which, in turn, required fixing many
unrelated things to get the noise level of the tests down enough to see
if anything got broken by the rewrite.

The only thing that appears to genuinely not work now after these
changes is the circular import part. In fact, circular imports are the
mean kid on the playground. They ruin everything. Removing support for
circular imports would make the rest of the code drastically simpler and
shorter.