Create temporary during cyclic import resolution
Before an import is fully resolved and can therefore be added as a
symbol to an environment, we instead add a temporary name referring to
a active binding that keeps track of where it’s being copied to during
initialization.

During the finalization of that import, we follow the trail of copies of
the active binding and replace them by the now finalized module
environment.