Fix errors in documentation
`R CMD CHECK` had multiple complaints about the documentation, some of
which were legitimate. To appease it, I’ve changed the format of the
`use` documentation to adhere strictly to the dictated convention, even
though it decreases its readability. We may want to revisit this in the
future, once the package has a first version on CRAN.