Replace the FFI vignette with something simpler
Using Rcpp for the FFI vignette makes sense because of its ease of use
and powerful API. Unfortunately, it also incurs a (quite heavy) added
dependency for building the package.

More seriously, though, is the fact that building the vignette
reproducibly deletes intermediate files that were copied somewhwere.
Unfortunately, this is a step in the Rcpp build workflow. And as a
consequence, `R CMD check` fails in the step that *rebuilds* the
vignettes.

The new vignette solves this by using simpler instructions for building
a C source file. It also moves the focus from FFI in general towards
compiled code in particular (FFI in R seems to be largely synomymous
with compiled code, but the terminology is vague and not widespread).