Make target of dispatched `help` call more general
When invoking `help` on a package, this gets dispatched to
`utils::help`. Except if some other attached environment overwrites this
function, we want to call *that* instead, please. So far, we had
hard-coded support for `devtools_shims`. This commit makes the search
more general, it should now find *any* function `help` that’s visible in
the caller’s scope, and pick the first of those according to R’s name
lookup rules (ignoring the one from this package).