Add missing vignette index to build bundle
The `build` directory was previously excluded from the R build process
because I used it for my own build tools, not realising that R needs it
to store the vignette index which should be included in source packages.

This change moves the custom build scripts to a different directory,
un-ignores `build`, and thus allows R CMD build to do its thing.