Add Makefile rule for a static vignette CRAN build
CRAN complains about modifications to “the user's global environment or
the user's home filespace" in the vignettes "by deleting objects". It’s
unclear what this refers to — possibly something caused by `R CMD
SHLIB`. At any rate, the CRAN rules for vignettes seem to be
restrictive, and in the interest of getting this package submitted, this
commit adds a rule which precompiles the vignettes and patches the
source files to refer to the precompiled versions. That way, CRAN should
be satisfied, since the vignettes no longer effect *any* changes to the
user system, while at the same time producing identical output (except
for the *.R vignette output files, which now consist entirely of
commented-out code. So be it).