Store sanitised tracebacks when rethrowing errors
`rethrow_on_error` previously caused the error traceback to diverge into
the that function rather than preserving the call stack down to the call
that caused the actual error. This commit use an `on.exit` hook in
`rethrow_on_error` to instead restore (a sanitised version of) the
original traceback, which is more informative since it contains the
actual error location.

This also makes the insertion of the actual error call into the error
message (“… (inside …)”) unnecessary, and this commit thus removes that.